Move FileSystemError definition to own file
