+++ pkg/system/file_system_funcs.go
@@ -12,32 +12,6 @@ import (
 	"github.com/setlog/trivrost/pkg/misc"
 )
 
-type FileSystemError struct {
-	message      string
-	causingError error
-}
-
-func (fse *FileSystemError) Error() string {
-	var causingErrorMessage string
-	if fse.causingError == nil {
-		causingErrorMessage = "<nil>"
-	} else {
-		causingErrorMessage = fse.causingError.Error()
-	}
-	if fse == nil {
-		return "<nil>: " + causingErrorMessage
-	}
-	return fse.message + ": " + causingErrorMessage
-}
-
-func (fse *FileSystemError) Unwrap() error {
-	return fse.causingError
-}
-
-func NewFileSystemError(message string, cause error) *FileSystemError {
-	return &FileSystemError{message: message, causingError: cause}
-}
-
 func MustMakeTempDirectory(forFolderAtPath string) string {
 	randomHex := misc.MustGetRandomHexString(8)
 	tempDirPath := filepath.Join(filepath.Dir(forFolderAtPath), "~"+filepath.Base(forFolderAtPath)+".dl."+randomHex)
